16-byte align round keys in AES-NI
See #70.

RKs are not explicitly loaded and stored so the pointers need to be aligned.

A better solution than allocating a little more and aligning on access would be
to make a utility to allocate them aligned in the first place. The best solution
would be to have a reusable one.