Fix overflow in Counter.add16.
I wrote in C since doing the addition with Int64.t and then an unsigned
comparison in ocaml adds a lot of clutter to the code.

As there was precedence in misc.c with caml_nc_count_16_be this seemed like the
right approach.

I'm implementing SSH and it uses the whole 128 bits as the counter in AES-CTR,
so I can expect this to overflow at some point, since the initial IV can be
anything.