Provide betole|letobe macros on OS X.
OS X endian.h doesn't define betole|letobe functions. It's also not optional
depending on a #define, it's really not there. This seems to be the common
solution people adopt.

This change adds all macros, even ones that are not used by nocrypto to match
endian.h on other systems, otherwise this problem will pop up again once another
betole/letobe function is used.

It was broken on the cpu_to_* conversion on commit:
commit 555d2fe99b32481801bbfeb346b0b87dc122c7c0
Author: pqwy <david@numm.org>
Date:   Sat Dec 9 21:41:17 2017 -0500

    uncrap counters